% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_SparsePVCM.R
\name{bic_SparsePVCM}
\alias{bic_SparsePVCM}
\title{Compute BIC value for SPVCM}
\usage{
bic_SparsePVCM(y, M, Theta_fit)
}
\arguments{
\item{y}{The vector of a response variable}

\item{M}{Design matrix for the vectorized Theta matrix}

\item{Theta_fit}{The output list of dys_pvcm function}
}
\value{
A list that contains BIC value and an index vector to locate nonzero coefficient functions
}
\description{
\code{bic_spvcm} computes the BIC value for a given SPVCM for model selection.
}
