% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dys_SparsePVCM.R
\name{dys_SparsePVCM}
\alias{dys_SparsePVCM}
\title{Davis-Yin splitting algorithm for SparsePVCM}
\usage{
dys_SparsePVCM(y, M, Theta_initial, w1, w2, rho, tau, max_iter, tol)
}
\arguments{
\item{y}{The vector of a response variable}

\item{M}{Design matrix for the vectorized Theta matrix}

\item{Theta_initial}{A matrix for an initial value}

\item{w1}{A numeric vector for the weights of nuclear norm}

\item{w2}{A numeric vector for the adaptive weights of group lasso}

\item{rho}{The step size of DYS algorithm}

\item{tau}{A vector of two tuning parameters for nuclear norm and group lasso}

\item{max_iter}{A numeric value for the maximum number of iterations}

\item{tol}{Tolerance value used as a stopping criterion}
}
\value{
A list that contains the results of DYS algorithm
}
\description{
\code{dys_SparsePVCM} is a function to find the optimal solution with given
two weight vectors and two tuning parameters.
}
