% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_SparsePVCM.R
\name{predict_SparsePVCM}
\alias{predict_SparsePVCM}
\title{Fitted values and estimated function evaluation}
\usage{
predict_SparsePVCM(fit, U_new, X_new = NULL)
}
\arguments{
\item{fit}{The output of SparsePVCM}

\item{U_new}{A numeric vector of index variable values to evaluate estimated functions}

\item{X_new}{A design matrix of new predictor values that corresponds to \code{U_new}}
}
\value{
A list that contains fitted response, evaluated coefficient functions and principal functions at given U_new
}
\description{
A function to evaluate the estimated functions at some particular values of index variable and
to calculate the fitted values of response if new predictor values are given
}
