% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Generate data for simulation studies}
\usage{
sim_data(snr = 5)
}
\arguments{
\item{snr}{Signal-to-noise ratio}
}
\value{
A list that contains all data with 3 different scenarios
}
\description{
Generate data for simulation studies
}
