% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparsePVCM.R
\name{SparsePVCM}
\alias{SparsePVCM}
\title{Fit a Sparse Principal Varying Coefficient Model}
\usage{
SparsePVCM(
  y_sample,
  X_sample,
  U_sample,
  K = 10,
  num_core = NULL,
  num_tun_1 = 10,
  num_tun_2 = 10,
  max_iter = 100,
  tol = 1e-05
)
}
\arguments{
\item{y_sample}{The vector of a response variable}

\item{X_sample}{The design  matrix of covariates}

\item{U_sample}{The vector of an index variable}

\item{K}{The number of cubic B-spline basis functions. If not given, the total number minus one will be used.}

\item{num_core}{The number of cores for parallel computing. If it is not given, the total number of cores minus 1 will be used.}

\item{max_iter}{A numeric value for the maximum number of iterations}

\item{tol}{Tolerance value used as a stopping criterion}

\item{num_tau_1}{The number of nonzero tuning parameter candidates for nuclear norm. Default is 6}

\item{num_tau_2}{The number of nonzero tuning parameter candidates for group lasso. Default is 6}
}
\value{
A list that contains the results of SPVCM estimation
}
\description{
\code{SparsePVCM} is a function to return the optimal SparsePVCM by choosing the one with the
smallest BIC value.
}
