% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCM.R
\name{VCM}
\alias{VCM}
\title{Fit a Varying Coefficient Model with a local polynomial regression}
\usage{
VCM(
  y_sample,
  X_sample,
  U_sample,
  index_i,
  U_new = NULL,
  num_core = NULL,
  num_fold = 5,
  num_h = 50
)
}
\arguments{
\item{y_sample}{The vector of a response variable}

\item{X_sample}{The design  matrix of covariates}

\item{U_sample}{The vector of an index variable}

\item{index_i}{The integer vector of subject index ranging from 1 to the total number of subjects. Each entry corresponds to that of \code{y_sample}}

\item{U_new}{A vector of any numeric values within the range of \code{U_sample} to evaluate coefficient functions at the given values as \code{beta_Unew}}

\item{num_core}{The number of cores for parallel computing. If it is not given, the total number of cores minus 1 will be used.}

\item{num_fold}{The number of folds to implement K-fold cross validation to choose the optimal bandwidth}

\item{num_h}{The number of candidates for the optimal bandwidth computed based on Fan and Gijbels 1995}
}
\value{
A list that contains the results of VCM estimation
}
\description{
\code{VCM} is a function to fit VCM with a local polynomial regression through Epanechnikov kernel.
}
